requirements.txt
streamlit
pandas
python-pptx
requests
deep-translator